Optimize the traveling thief problem (TTP)  with Nature-Inspired Algorithms
The traveling thief problem (TTP) is a complex problem that combines two well-known NP-hard problems, Traveling Salesman Problem (TSP) and Knapsack Problem (KNP). The problem involves finding the shortest route that visits a set of cities. Each city has a collection of items. The goal is to select a subset of items to maximize profit while considering the weight of the items and the speed of the thief. The problem is relevant for many real-world applications, such as logistics, delivery, transportation, and planning.